# docs/gui/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchBenchmarkRequest": {
        "properties": {
          "suite_name": {
            "default": "quick_test",
            "enum": [
              "quick_test",
              "appendix_32x32",
              "commitment_types",
              "main_matrix"
            ],
            "title": "Suite Name",
            "type": "string"
          }
        },
        "title": "BatchBenchmarkRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SimulationRequest": {
        "description": "Specification for single or comparative simulation runs.",
        "properties": {
          "agent_count": {
            "default": 10,
            "maximum": 100.0,
            "minimum": 2.0,
            "title": "Agent Count",
            "type": "integer"
          },
          "commitment_type": {
            "default": "SC",
            "enum": [
              "SC",
              "DC",
              "ZC"
            ],
            "title": "Commitment Type",
            "type": "string"
          },
          "compare_commitment_type": {
            "anyOf": [
              {
                "enum": [
                  "SC",
                  "DC",
                  "ZC"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compare Commitment Type"
          },
          "compare_solver": {
            "anyOf": [
              {
                "pattern": "^(EECBS-1.1|EECBS-1.0|CBS|Prioritized|PathAware|HeatMap|Greedy|Conceder)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compare Solver"
          },
          "fov_size": {
            "default": 5,
            "maximum": 15.0,
            "minimum": 3.0,
            "title": "Fov Size",
            "type": "integer"
          },
          "grid_height": {
            "default": 16,
            "maximum": 64.0,
            "minimum": 4.0,
            "title": "Grid Height",
            "type": "integer"
          },
          "grid_width": {
            "default": 16,
            "maximum": 64.0,
            "minimum": 4.0,
            "title": "Grid Width",
            "type": "integer"
          },
          "initial_tokens": {
            "default": 5,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Initial Tokens",
            "type": "integer"
          },
          "max_steps": {
            "default": 100,
            "maximum": 300.0,
            "minimum": 10.0,
            "title": "Max Steps",
            "type": "integer"
          },
          "obstacle_density": {
            "default": 0.0,
            "maximum": 0.4,
            "minimum": 0.0,
            "title": "Obstacle Density",
            "type": "number"
          },
          "random_seed": {
            "default": 42,
            "title": "Random Seed",
            "type": "integer"
          },
          "setting": {
            "default": 4,
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Setting",
            "type": "integer"
          },
          "solver": {
            "default": "PathAware",
            "pattern": "^(EECBS-1.1|EECBS-1.0|CBS|Prioritized|PathAware|HeatMap|Greedy|Conceder)$",
            "title": "Solver",
            "type": "string"
          },
          "timeout_sec": {
            "default": 60.0,
            "maximum": 600.0,
            "minimum": 1.0,
            "title": "Timeout Sec",
            "type": "number"
          }
        },
        "title": "SimulationRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Decentralized MAPF simulation and centralized solver comparisons, independent trajectory validation, replay and experiment analysis.",
    "title": "DEC-MAPF - Research Workspace",
    "version": "0.1.0a1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/benchmark/cancel/{task_id}": {
      "post": {
        "operationId": "cancel_benchmark_api_benchmark_cancel__task_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Benchmark Api Benchmark Cancel  Task Id  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Benchmark"
      }
    },
    "/api/benchmark/start": {
      "post": {
        "operationId": "start_benchmark_api_benchmark_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchBenchmarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Start Benchmark Api Benchmark Start Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Benchmark"
      }
    },
    "/api/benchmark/status/{task_id}": {
      "get": {
        "operationId": "get_benchmark_status_api_benchmark_status__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Benchmark Status Api Benchmark Status  Task Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Benchmark Status"
      }
    },
    "/api/benchmark/stream/{task_id}": {
      "get": {
        "description": "SSE stream endpoint supporting Last-Event-ID header and cursor query.",
        "operationId": "stream_benchmark_events_api_benchmark_stream__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Benchmark Events"
      }
    },
...
    "/api/export/latex/{suite_name}": {
      "get": {
        "operationId": "export_latex_table_api_export_latex__suite_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "suite_name",
            "required": true,
            "schema": {
              "title": "Suite Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Export Latex Table Api Export Latex  Suite Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Latex Table"
      }
    },
...
    "/api/results/{suite_name}": {
      "get": {
        "operationId": "get_archived_results_api_results__suite_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "suite_name",
            "required": true,
            "schema": {
              "title": "Suite Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Archived Results Api Results  Suite Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Archived Results"
      }
    },
    "/api/simulate": {
      "post": {
        "operationId": "run_simulation_api_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Simulation"
      }
    },
...
  }
}